client-ui/requirements.txt
blinker==1.7.0
boto3==1.34.82
botocore==1.34.82
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cognitojwt==1.4.1
cryptography==42.0.5
ecdsa==0.19.0
Flask==2.1.3
Flask-Cors==4.0.0
flask-dynamo==0.1.2
flask_cognito_lib==1.6.2
gunicorn
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
tzdata==2024.1
urllib3==1.26.18
Werkzeug==2.2.2
zipp==3.18.1
